Simplify Remote MCP example Hono app (#27)
The goal of this demo app is to serve as a simplified example that's easy to follow.

The code for the example was largely generated by Claude and spread the app code across a bunch of files and interleaved static html content with the authentication logic. It was pretty hard to follow.

This puts all of the logic for the Hono app into one file, makes it clear where you would add your own user / password validation logic, and hides away all of the incidental HTML generation in the utils.ts file.